DEV-26710: Remove OHEM layer from Curalate Caffe https://curalate.atlassian.net/browse/DEV-26710
We are having an issue where the `box_annotator_ohem_layer` does not compile properly when building for the GPU.  For the sake of speed, we are going to remove this layer for now since it is not needed for deployment.  We will need this layer to train an OHEM Detection Model so we should revisit adding this layer back in the future.